+++ thewalrus/tests/test_hafnian_approx.py
@@ -29,7 +29,7 @@ def test_rank_one(n):
     x = np.random.rand(n)
     A = np.outer(x, x)
     exact = factorial2(n - 1) * np.prod(x)
-    approx = hafnian(A, approx=True, num_samples=10000)
+    approx = hafnian(A, method="barvinok", num_samples=10000)
     assert np.allclose(approx, exact, rtol=2e-1, atol=0)
 
 
@@ -37,7 +37,7 @@ def test_approx_complex_error():
     """Check exception raised if matrix is complex"""
     A = 1j * np.ones([6, 6])
     with pytest.raises(ValueError, match="Input matrix must be real"):
-        hafnian(A, approx=True)
+        hafnian(A, method="barvinok")
 
 
 def test_approx_negative_error():
@@ -45,13 +45,13 @@ def test_approx_negative_error():
     A = np.ones([6, 6])
     A[0, 0] = -1
     with pytest.raises(ValueError, match="Input matrix must not have negative entries"):
-        hafnian(A, approx=True)
+        hafnian(A, method="barvinok")
 
 
 @pytest.mark.parametrize("n", [6, 8])
 def test_ones_approx(n):
     """Check hafnian_approx(J_2n)=(2n)!/(n!2^n)"""
     A = np.float64(np.ones([2 * n, 2 * n]))
-    haf = hafnian(A, approx=True, num_samples=10000)
+    haf = hafnian(A, method="barvinok", num_samples=10000)
     expected = fac(2 * n) / (fac(n) * (2**n))
     assert np.abs(haf - expected) / expected < 0.2