+++ thewalrus/quantum/photon_number_distributions.py
@@ -14,6 +14,7 @@
 """
 Functions for calculating the photon number distributions of various states.
 """
+
 import numpy as np
 from scipy.stats import nbinom
 from scipy.special import binom, hyp2f1