+++ thewalrus/tests/test_samples.py
@@ -13,9 +13,10 @@
 # limitations under the License.
 """Tests for the hafnian sampling functions"""
 # pylint: disable=no-self-use,redefined-outer-name
-import pytest
+import sys
 
 import numpy as np
+import pytest
 from scipy.stats import nbinom
 
 from thewalrus.samples import (
@@ -102,10 +103,10 @@ class TestHafnianSampling:
 
         n_cut = 10
         samples = hafnian_sample_state(sigma, samples=n_samples, cutoff=n_cut)
-        bins = np.arange(0, max(samples) + 1, 1)
+        bins = np.arange(0, np.max(samples) + 1, 1)
         (freq, _) = np.histogram(samples, bins=bins)
         rel_freq = freq / n_samples
-        nm = max(samples) // 2
+        nm = np.max(samples) // 2
 
         x = nbinom.pmf(np.arange(0, nm, 1), 0.5, np.tanh(np.arcsinh(np.sqrt(mean_n))) ** 2)
         x2 = np.zeros(2 * len(x))
@@ -233,6 +234,10 @@ class TestHafnianSampling:
         assert np.allclose(mean_n, approx_mean_n, rtol=2e-1)
 
     @pytest.mark.parametrize("parallel", [True, False])
+    @pytest.mark.skipif(
+        (lambda parallel: parallel is True) and (sys.platform == "darwin"),
+        reason="does not run on macos",
+    )
     def test_single_pm_graphs(self, parallel):
         """Tests that the number of photons is the same for modes i and n-i
         in the special case of a graph with one single perfect matching
@@ -309,7 +314,7 @@ class TestHafnianSampling:
         mean_n = 0.5
         sigma = (2 * mean_n + 1) * np.identity(2)
         samples = sample_func(sigma, samples=n_samples)
-        bins = np.arange(0, max(samples), 1)
+        bins = np.arange(0, np.max(samples), 1)
         (freq, _) = np.histogram(samples, bins=bins)
         rel_freq = freq / n_samples
 
@@ -445,6 +450,10 @@ class TestTorontonianSampling:
         assert np.all(np.abs(rel_freq - probs) < rel_tol / np.sqrt(n_samples))
 
     @pytest.mark.parametrize("parallel", [True, False])
+    @pytest.mark.skipif(
+        (lambda parallel: parallel is True) and (sys.platform == "darwin"),
+        reason="does not run on macos",
+    )
     def test_torontonian_sample_graph(self, parallel):
         """Test torontonian sampling from a graph"""
         A = np.array([[0, 3.0 + 4j], [3.0 + 4j, 0]])