+++ .github/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### Bug fixes
 
+* Add the calculation method of `takagi` when the matrix is diagonal. [(#394)](https://github.com/XanaduAI/thewalrus/pull/394)
+
 ### Documentation
 
 ### Contributors