+++ .github/CHANGELOG.md
@@ -17,6 +17,8 @@
 
 * Add the calculation method of `takagi` when the matrix is diagonal. [(#394)](https://github.com/XanaduAI/thewalrus/pull/394)
 
+* Add the lines for avoiding the comparison of np.ndarray and list. 
+
 ### Documentation
 
 ### Contributors